+++ src/common/tokenHelpers.ts
@@ -71,8 +71,8 @@ export const validateToken = (token: EdgeToken): void => {
 
 /**
  * Validates a currency code.
- * Some weird but valid examples include: T, BUSD.e, xBOO, 1INCH, BADGER
+ * Some weird but valid examples include: T, BUSD.e, xBOO, 1INCH, BADGER, BSC-USD
  */
-const isCurrencyCode = (code: string): boolean => {
-  return /^[.a-zA-Z0-9]+$/.test(code)
+export const isCurrencyCode = (code: string): boolean => {
+  return /^[a-zA-Z0-9]+([.-][a-zA-Z0-9]+)?$/.test(code)
 }