+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added: Arbitrum One network support
 - added: Base network support
 - added: Bitstamp EUR Ripple token
+- added: BSC Tokens BSC-USD,DOGE,BTCB,ETH,TUSD,DOT,WBNB,LINK, and MATIC
 - removed: Unused `EthereumNetworkInfo` params
 
 ## 3.0.4 (2024-01-30)